created a separated file for checksum
